requirements.txt
viam-sdk
numpy
opencv-python
pillow